+++ src/main/java/com/app/smartdrive/api/entities/master/TemplateTaskWorkOrder.java
@@ -14,13 +14,13 @@ public class TemplateTaskWorkOrder {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "tewo_id", updatable = false, nullable = false)
-    private int tewoId;
+    private Long tewoId;
 
     @Column(name = "tewo_name")
     private String tewoName;
 
     @Column(name = "tewo_testa_id", insertable = false, nullable = false)
-    private int tewoTestaId;
+    private Long tewoTestaId;
 
     @ManyToOne
     @JoinColumn(name = "tewo_testa_id", insertable = false, updatable = false)
